Batch exception fix when unwrapping exception Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
