+++ server/implementation/src/main/java/io/smallrye/graphql/execution/error/ExceptionHandler.java
@@ -2,6 +2,10 @@ package io.smallrye.graphql.execution.error;
 
 import static io.smallrye.graphql.SmallRyeGraphQLServerLogging.log;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CompletionException;
+
 import graphql.ExceptionWhileDataFetching;
 import graphql.execution.DataFetcherExceptionHandler;
 import graphql.execution.DataFetcherExceptionHandlerParameters;
@@ -9,6 +13,7 @@ import graphql.execution.DataFetcherExceptionHandlerResult;
 import graphql.execution.ResultPath;
 import graphql.language.SourceLocation;
 import io.smallrye.graphql.bootstrap.Config;
+import io.smallrye.graphql.execution.datafetcher.DataFetcherException;
 
 /**
  * Here we have the ability to mask certain messages to the client (for security reasons)
@@ -19,15 +24,21 @@ public class ExceptionHandler implements DataFetcherExceptionHandler {
 
     private final Config config;
     private final ExceptionLists exceptionLists;
+    protected List<String> unwrapExceptions = new ArrayList<>();
 
     public ExceptionHandler(Config config) {
         this.config = config;
         this.exceptionLists = new ExceptionLists(config.getHideErrorMessageList(), config.getShowErrorMessageList());
+        if (config.getUnwrapExceptions().isPresent()) {
+            this.unwrapExceptions.addAll(config.getUnwrapExceptions().get());
+        }
+        this.unwrapExceptions.addAll(DEFAULT_EXCEPTION_UNWRAP);
     }
 
     @Override
     public DataFetcherExceptionHandlerResult onException(DataFetcherExceptionHandlerParameters handlerParameters) {
         Throwable throwable = handlerParameters.getException();
+        throwable = unwrapThrowable(throwable);
         SourceLocation sourceLocation = handlerParameters.getSourceLocation();
         ResultPath path = handlerParameters.getPath();
         ExceptionWhileDataFetching error = getExceptionWhileDataFetching(throwable, sourceLocation, path);
@@ -56,5 +67,26 @@ public class ExceptionHandler implements DataFetcherExceptionHandler {
                 return new GraphQLExceptionWhileDataFetching(path, throwable, sourceLocation);
             }
         }
+    }
+
+    private Throwable unwrapThrowable(Throwable t) {
+        if (shouldUnwrapThrowable(t)) {
+            t = t.getCause();
+            return unwrapThrowable(t);
+        }
+        return t;
+    }
+
+    private boolean shouldUnwrapThrowable(Throwable t) {
+        return unwrapExceptions.contains(t.getClass().getName()) && t.getCause() != null;
+    }
+
+    private static final List<String> DEFAULT_EXCEPTION_UNWRAP = new ArrayList<>();
+
+    static {
+        DEFAULT_EXCEPTION_UNWRAP.add(CompletionException.class.getName());
+        DEFAULT_EXCEPTION_UNWRAP.add(DataFetcherException.class.getName());
+        DEFAULT_EXCEPTION_UNWRAP.add("javax.ejb.EJBException");
+        DEFAULT_EXCEPTION_UNWRAP.add("jakarta.ejb.EJBException");
     }
 }
