+++ server/tck/src/test/resources/tests/batch/error/input.graphql
@@ -1,0 +1,8 @@
+{
+  batchPojos{
+    business
+    runtime
+    businesses
+    runtimes
+  }
+}
