+++ server/tck/src/test/resources/tests/batch/normal/input.graphql
