+++ server/tck/src/test/resources/tests/batch/normal/test.properties
@@ -1,0 +1,2 @@
+ignore=false
+priority=100
